Copied contents from Slider.jsx to Slider-fix.jsx
